backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9
supabase==2.3.4
python-dotenv==1.0.1
google-api-python-client==2.122.0
groq==0.4.2
openai==1.14.0
markdown==3.6
PyJWT==2.8.0
bcrypt==5.0.0
